requirements.txt
launchdarkly-server-sdk>=9.0.0
launchdarkly-server-sdk-ai>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.0.20
python-dotenv>=1.0.0
